Further cleanups in the split FileUtil
